requirements.txt
colorama==0.4.4
python-dotenv==1.0.0
